Support NAT-PMP, at least on Linux.
Other OSes require writing a `getGateway()` implementation.  The BSDs
looks trivial, but I need to spin up a VM to do so.  No idea how this
will work with Windows, I''l deal with that when I need to.

NB: NAT-PMP's "-v" output is non-existent.  It will be added to shortly.